MD: raid5 avoid unnecessary zero page for trim
We want to avoid zero discarded dev page, because it's useless for
discard.  But if we don't zero it, another read/write hit such page in
the cache and will get inconsistent data. To avoid zero the page, we
set R5_WantZeroFill for discarded dev page. Every time before the page
is accessed and the flag is set, we zero the page and clear the
flag. If the page will be drained or computed, we just clear the flag
for it. In this way, the dev page data is alway consistent. And since
the chance discarded data is accessed soon is low, zero discard dev
page is largely avoided.

Signed-off-by: Shaohua Li <shli@fusionio.com>
Signed-off-by: NeilBrown <neilb@suse.de>