include/linux/init.h: use the stringify operator for the __define_initcall macro
Currently the __define_initcall macro takes three arguments, fn, id and
level.  The level argument is exactly the same as the id argument but
wrapped in quotes.  To overcome this need to specify three arguments to
the __define_initcall macro, where one argument is the stringification of
another, we can just use the stringification macro instead.

Signed-off-by: Matthew Leach <matthew@mattleach.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>