include/linux/kfifo.h: replace open-coded type check code with typecheck()
In the kfifo macros, one piece of code which is arounded by if(0) will
check the type of __tmp->ptr_const and __buf.  If they are of different
types a compilation warning is generated..  This piece of code is not self
explaining and a little bit hard to understand.

Based on Andrew Morton's suggestion, replace this with typecheck() which
will be easy to understand.

Signed-off-by: Wei Yang <weiyang@linux.vnet.ibm.com>
Cc: richard -rw- weinberger <richard.weinberger@gmail.com>
Acked-by: Stefani Seibold <stefani@seibold.net>
Cc: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>